requirements.txt
flask
python-dotenv
flask-sqlalchemy
flask-migrate
flask-script
flask-restplus
mysqlclient
